rename icon.constants to gen
