+++ shared/common-adapters/icon.desktop.tsx
@@ -3,7 +3,7 @@ import * as Styles from '../styles'
 import logger from '../logger'
 import React, {Component} from 'react'
 import shallowEqual from 'shallowequal'
-import {iconMeta} from './icon.constants'
+import {iconMeta} from './icon.constants-gen'
 import {resolveImageAsURL} from '../desktop/app/resolve-root.desktop'
 import {invert} from 'lodash-es'
 import {Props, IconType} from './icon'
