+++ shared/common-adapters/icon.shared.tsx
@@ -1,6 +1,6 @@
 import * as Styles from '../styles'
 import {IconType, SizeType} from './icon'
-import {iconMeta} from './icon.constants'
+import {iconMeta} from './icon.constants-gen'
 import './icon.css'
 
 export function defaultColor(type: IconType): string | null {
