+++ shared/constants/types/people.tsx
@@ -1,6 +1,6 @@
 import * as I from 'immutable'
 import * as RPCTypes from './rpc-gen'
-import {IconType} from '../../common-adapters/icon.constants'
+import {IconType} from '../../common-adapters/icon.constants-gen'
 
 export type ItemTypeEnum = RPCTypes.HomeScreenItemType
 export type ItemType = keyof typeof RPCTypes.HomeScreenItemType
