+++ shared/people/todo/container.tsx
@@ -4,7 +4,7 @@ import * as PeopleGen from '../../actions/people-gen'
 import * as Types from '../../constants/types/people'
 import * as Tabs from '../../constants/tabs'
 import * as SettingsTabs from '../../constants/settings'
-import {IconType} from '../../common-adapters/icon.constants'
+import {IconType} from '../../common-adapters/icon.constants-gen'
 import {todoTypes} from '../../constants/people'
 import {connect, isMobile} from '../../util/container'
 import * as Tracker2Gen from '../../actions/tracker2-gen'
