+++ shared/router-v2/router.native.tsx
@@ -1,5 +1,5 @@
 import * as Kb from '../common-adapters/mobile.native'
-import {IconType} from '../common-adapters/icon.constants'
+import {IconType} from '../common-adapters/icon.constants-gen'
 import * as Styles from '../styles'
 import * as React from 'react'
 import GlobalError from '../app/global-errors/container'
