+++ shared/wallets/airdrop/index.tsx
@@ -3,7 +3,7 @@ import * as Kb from '../../common-adapters'
 import * as Constants from '../../constants/wallets'
 import * as Styles from '../../styles'
 import * as Types from '../../constants/types/wallets'
-import {iconMeta} from '../../common-adapters/icon.constants'
+import {iconMeta} from '../../common-adapters/icon.constants-gen'
 import openURL from '../../util/open-url'
 
 export type Props = {
